+++ tests/test_image/test_iterm2.py
@@ -414,8 +414,8 @@ class TestRenderLines:
             if scale == 0.0:
                 continue
             self.trans.scale = scale
-            if 0 in self.trans.rendered_size:
-                continue
+
+            assert 0 not in self.trans.rendered_size
             for ITerm2Image._TERM in supported_terminals:
                 self._test_image_size(self.trans, term=ITerm2Image._TERM)
 
@@ -699,8 +699,8 @@ class TestRenderWhole:
             if scale == 0.0:
                 continue
             self.trans.scale = scale
-            if 0 in self.trans.rendered_size:
-                continue
+
+            assert 0 not in self.trans.rendered_size
             for ITerm2Image._TERM in supported_terminals:
                 self._test_image_size(self.trans, term=ITerm2Image._TERM)